Moved useReducerWithHistory to its own file.
