+++ src/index.js
@@ -1,5 +1,6 @@
-import React, { useEffect, Fragment, createContext, useContext, useRef, useState } from 'react';
+import React, { useEffect, Fragment, createContext, useContext } from 'react';
 import { FaUndo, FaRedo } from "react-icons/fa"
+import useReducerWithHistory from "./useReducerWithHistory"
 import ReactDOM from 'react-dom';
 import './index.css';
 
@@ -8,46 +9,6 @@ let gameState = null
 
 // Initialize a context for the game
 const gameContext = createContext()
-
-// Custom hook for adding history functions to a reducer pattern.
-function useReducerWithHistory(reducer, state) {
-  // Use a reference for persistent history
-  const history = useRef([state])
-
-  // Set some state for the current index
-  const [index, setIndex] = useState(0)
-
-  // Function to determin if undo is possible
-  function canUndo() {
-    return (index > 0)
-  }
-
-  // Function to rewind index by 1
-  function undo() {
-    setIndex(canUndo() ? index - 1 : index)
-  }
-
-  // Function to determine if redo is possible
-  function canRedo() {
-    return (index < history.current.length - 1)
-  }
-
-  // Function to increase index by 1
-  function redo() {
-    setIndex(canRedo() ? index + 1 : index)
-  }
-
-  // Dispatcher that preserves history when calling the reducer
-  function dispatch(action) {
-    const newState = reducer(history.current[index], action)
-    history.current = history.current.slice(0, index + 1)
-    history.current.push(newState)
-    setIndex(history.current.length - 1)
-  }
-
-  // Return the current state, and the new functions
-  return [history.current[index], dispatch, canUndo, undo, canRedo, redo]
-}
 
 function Game() {
   // Create the reducer pattern for the state
